Optimizations

+++ include/functions.h
@@ -8,11 +8,12 @@
 #ifndef INCLUDE_FUNCTIONS_H_
 #define INCLUDE_FUNCTIONS_H_
 
+#include <stddef.h>
 #include "asmop.h"
 #include "tokens.h"
-#include "../include/types.h"
+#include "types.h"
 
-typedef ASMOP* ASM_FUNC_GENERATE(Token*, ASMOP*, int*, int*);
+typedef ASMOP* ASM_FUNC_GENERATE(Token*, ASMOP*, size_t*, int*);
 
 typedef struct {
 	char name[64];
@@ -24,6 +25,6 @@ typedef struct {
 void GetFunction(const char *name, Function *f, int *found);
 
 void Init_Functions(char* (*varname)(const char *identifier_name),
-		TYPE (*value_instructions)(Token*, int, ASMOP*, int*, int*));
+		TYPE (*value_instructions)(Token*, int, ASMOP*, size_t*, int*));
 
 #endif /* INCLUDE_FUNCTIONS_H_ */
